Had to move all commands to one master file
It turns out that the reason why many commands weren't being loaded in
by click was because the python files themselves were not being run. I
think this can be amended by added the commands file as a package in
setup.py, or by importing in the files forcing them to be run. The
problem with the latter is a fear of circular imports now or in the
future.