Added the create rebalance command
The purpose of the rebalance command isn't to run a rebalance ( although
a --run option will be added in the future ). The purpose is to define
the rebalance that will be attempted and save that copy because, more
than likely, you will need to rerun the exact same rebalance route.